LayoutTests/plugins/netscape-plugin-setwindow-size.html

<html>
<head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<body style="margin:0px;overflow:hidden">
<embed id="testPlugin" type="application/x-webkit-test-netscape" logfirstsetwindow="true" width="100%" height="200"></embed>
<div>
    This tests that a plug-in with a percentage width gets a correct NPP_SetWindow the first time.
</div>
<body>
</html>
