LayoutTests/plugins/netscape-plugin-setwindow-size-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: NPP_SetWindow: 800 200

This tests that a plug-in with a percentage width gets a correct NPP_SetWindow the first time.
